Rename variables to increase consistency
The superclass that encodes the logic for a specific family of methods
in the NaCl bindings is referred to as a delegate in other areas, so
that name is now used consistently when it is stored in a
variable. Likewise, subclasses that inherit from the delegate are also
consistently assigned to appropriately-named variables.