Rework automated testing setup
New features in this branch require an unreleased version of
libsodium. In the process of tweaking the build setup to use the latest
version on `master`, it made sense to *always* test against a local
build of libsodium.

This commit also cleans up the Rake tasks involved somewhat, and removes
a warning caused when building against newer versions of MiniTest.